Move holder test methods
